refactor: move results view in separate files
