+++ apps/votes_results/views/__init__.py
@@ -2,3 +2,5 @@ from .single_option_vote_view import *
 from .majority_judgment_vote_view import *
 from .votable_polls_view import *
 from .generic_vote_view import *
+from .results.single_option_results_view import *
+from .results.majority_judgment_results_view import *
