+++ apps/votes_results/views/single_option_vote_view.py
@@ -230,48 +230,3 @@ def single_option_recap_view(request: HttpRequest, poll_id: int):
                                                                       'mj_vote_counter': mj_vote_counter})
 
 
-def single_option_results_view(request: HttpRequest, poll_id: int):
-    """Render page with results.
-    Args:
-        request (HttpRequest): Request object.
-        poll_id (int): The poll id.
-    Returns:
-        HttpResponse: Rendered results page.
-        HttpResponseServerError: If DB is not initialized.
-    """
-
-    # if poll type is majority, we need to redirect 
-    # to majority results page
-    try:
-        # Retrieve poll
-        poll: PollModel = PollService.get_poll_by_id(poll_id)
-    except Exception:
-        raise Http404(f"Poll with id {poll_id} not found.")
-
-    # redirect to details page if poll is not yet open
-    if not poll.is_open():
-        return HttpResponseRedirect(reverse('apps.polls_management:poll_details', args=(poll_id,)))
-
-    if poll.poll_type == PollModel.PollType.MAJORITY_JUDJMENT:
-        return HttpResponseRedirect(reverse('apps.votes_results:majority_judgment_results', args=(poll_id,)))
-
-    # regular results page 
-    try:
-        poll_results: PollResult = SingleOptionVoteService.calculate_result(poll_id)
-    except PollDoesNotExistException:
-        raise Http404
-    except Exception:
-        # Internal error: you should inizialize DB first (error 500)
-        return HttpResponseServerError("Something got (slighly) terribly wrong. Please contact developers")
-
-    return render(request, 'votes_results/single_option_results.html', 
-        {'poll_results': poll_results,
-         'poll_info_to_view': {
-            'name': False,
-            'question': False,
-            'choices': False,
-            'status': True,
-            'datetimes': True,
-        }
-         }
-        )
